refactor(core): 清理 Agent Core 旧流水线死代码
确认 ReactEngine::execute_turn 为唯一执行入口后，移除从未被主链路调用的
plan-then-execute 旧流水线残留，共删除约 2500 行死代码。

删除（经全工作区引用追踪确认零调用）：
- execution/ 整个子树（plan_runner/step_executor/result_analyzer/verify/
  message/types）：顶层入口 execute_plan_with_execution_agent 无任何调用方
- agents/planning_agent.rs：公开导出但零外部调用
- agents/response_agent.rs：仅 VerifyExecutionRecord 被复用，prompt 逻辑零调用
- agents/execution_completion_agent.rs、execution_prompt_agent.rs：仅被
  死掉的 step_executor 引用
- orchestrator/ 整个目录：QueryOrchestrator 零调用，QueryMode 为永远返回
  MultiStepExecution 的 stub

迁移（保持对外 API 不变）：
- VerifyExecutionRecord 结构体定义收敛到 runtime.rs（本就从该处 re-export）
- QueryMode 枚举内联到 context/assembler.rs（消除对 orchestrator 的依赖）

附带清理：
- 移除 execution_mcp_agent 中因 step_executor 退场而变为死代码的
  build_internal_tool_error_result

行为无任何变化，全部 208 个测试通过。
