+++ crates/tiangong-core/src/lib.rs
@@ -10,7 +10,6 @@ pub mod context;
 pub mod core;
 pub mod core_config;
 pub mod event;
-pub mod execution;
 pub mod index;
 pub mod mcp;
 pub mod media;
@@ -19,7 +18,6 @@ pub mod memory;
 pub mod model;
 pub mod models_config;
 pub mod observe;
-pub mod orchestrator;
 pub mod permission;
 pub mod planner;
 pub mod plugin;
